+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Make cmd+s save forms automatically
 - Automatically save notes on change to match other inputs
+- automatically save notes when closing modal
+- fix several bugs when editing notes
 
 ## v8.0.0